Add config option precedence logic
Both the command line and the config file options are parsed.
Options from the command line have precedence over the same options from
the config file.

It should not be that difficult to add options from environment
variables to the mix as well (see the loop on lines 79-83 for the actual
priorization logic) when we want to allow configuration through those as
well.
Since containers are (usually) configured through environment variables,
that might be an interesting use case to add that.