Prepend / before message text if it's not there
We want the user to be able to type commands with or without leading
slash.
For example, 'version' and '/version' should both work.

Right now, a slash is prepended to all message text (unless the text
already starts with a slash).

This might be problematic later when we add state to the bot command
structure, but that's a problem to solve when we get there.
When we implement state, we'll probably just have to prepend a / to the
users' commands when they're in the top level state.