Don't accept commands from unknown users
Since the bot is a user interface on our Zabbix installation, we don't
want to execute commands unless the user sending the commands is known
in our Zabbix.

This is achieved by fetching a list of users from Zabbix that have the
Telegram mediatype on their account.

A bonus is that we can resolve the user ID of the incoming Telegram
messages to the corresponding Zabbix user and for example look up their
name.